retrieval-system/requirements.txt
openai>=1.50.0
pandas>=2.0.0
python-dotenv>=1.0.0
